Split `RegisterableSmartAnswers` from `PanopticonRegisterer`
By splitting this into two classes we can test `PanopticonRegisterer`
more robustly and re-use `RegisterableSmartAnswers` later, for sending
things to the content-store.
