+++ app/services/panopticon_registerer.rb
@@ -1,4 +1,10 @@
 class PanopticonRegisterer
+  attr_reader :unique_registerables
+
+  def initialize(unique_registerables)
+    @unique_registerables = unique_registerables
+  end
+
   def register
     require 'gds_api/panopticon'
     logger = GdsApi::Base.logger = Logger.new(STDERR).tap { |l| l.level = Logger::INFO }
@@ -11,26 +17,5 @@ class PanopticonRegisterer
       puts "Registering flow: #{registerable.slug} => #{registerable.title}"
       registerer.register(registerable)
     }
-  end
-
-private
-
-  def unique_registerables
-    # Picks smartdown of smart_answer for any dupe keys, same as routing behaviour
-    smart_answer_registrables.merge(smartdown_registrables).values
-  end
-
-  def smart_answer_registrables
-    flow_registry = SmartAnswer::FlowRegistry.new(FLOW_REGISTRY_OPTIONS)
-
-    Hash[flow_registry.flows.collect { |flow|
-      [flow.name, FlowRegistrationPresenter.new(flow)]
-    }]
-  end
-
-  def smartdown_registrables
-    Hash[SmartdownAdapter::Registry.instance.flows.collect { |flow|
-      [flow.name, SmartdownAdapter::FlowRegistrationPresenter.new(flow)]
-    }]
   end
 end
