+++ lib/tasks/panopticon.rake
@@ -3,6 +3,7 @@ require 'ostruct'
 namespace :panopticon do
   desc "Register application metadata with panopticon"
   task register: :environment do
-    PanopticonRegisterer.new.register
+    registerables = RegisterableSmartAnswers.new.unique_registerables
+    PanopticonRegisterer.new(registerables).register
   end
 end
