+++ test/unit/services/panopticon_registerer_test.rb
@@ -3,9 +3,10 @@ require 'test_helper'
 class PanopticonRegistererTest < ActiveSupport::TestCase
   def test_sending_item_to_panopticon
     request = stub_request(:put, %r[http://panopticon.dev.gov.uk/])
+    registerables = [OpenStruct.new, OpenStruct.new]
 
-    PanopticonRegisterer.new.register
+    PanopticonRegisterer.new(registerables).register
 
-    assert_requested(request, times: 42)
+    assert_requested(request, times: 2)
   end
 end
